+++ src/e3_dynamic_forms/conf.py
@@ -58,6 +58,12 @@ def get_form_schema_model():
     return apps.get_model(model_label)
 
 
+def get_form_response_model():
+    from django.apps import apps
+    model_label = getattr(settings, 'DYNAMIC_FORMS_RESPONSE_MODEL', 'e3_dynamic_forms.FormResponse')
+    return apps.get_model(model_label)
+
+
 def get_attachment_model():
     from django.apps import apps
     model_label = getattr(settings, 'DYNAMIC_FORMS_ATTACHMENT_MODEL', 'e3_dynamic_forms.Attachment')